7 Oct
Usual commit

+++ src/main/java/com/nagarro/fotonet/dao/UserDao.java
@@ -16,5 +16,5 @@ GenericDao<User,Integer> {
      * @param subscriptionId
      * @return
      */
-    User getUser(long subscriptionId) throws NonUniqueResultException;
+    User getUser(Integer subscriptionId) throws NonUniqueResultException;
 }
