+++ src/main/java/com/nagarro/fotonet/services/UserService.java
@@ -17,7 +17,7 @@ public interface UserService {
      * @param userName User name
      * @return
      */
-    User getUserBySubscriptionId(Long subscriptionId);
+    User getUserBySubscriptionId(Integer subscriptionId);
     
     
     /**
